requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt setup.py
#
aspy.yaml==1.1.2          # via pre-commit
cfgv==1.4.0               # via pre-commit
gitdb2==2.0.5             # via gitpython
gitpython==2.1.11
identify==1.2.1           # via pre-commit
importlib-metadata==0.8   # via pre-commit
importlib-resources==1.0.2  # via pre-commit
mypy-extensions==0.4.1    # via mypy
mypy==0.660
nodeenv==1.3.3            # via pre-commit
pre-commit==1.14.2
pyyaml==3.13              # via aspy.yaml, pre-commit
six==1.12.0               # via cfgv, pre-commit
smmap2==2.0.5             # via gitdb2
toml==0.10.0              # via pre-commit
typed-ast==1.2.0          # via mypy
virtualenv==16.3.0        # via pre-commit
zipp==0.3.3               # via importlib-metadata
